+++ source/timelinelib/config/dotfile.py
@@ -74,6 +74,7 @@ FUZZY_ICON = "fuzzy_icon"
 LOCKED_ICON = "locked_icon"
 HYPERLINK_ICON = "hyperlink_icon"
 DATE_FORMAT = "date_format"
+VERTCAL_SPACE_BETWEEN_EVENTS = "vertical_space_between_events"
 DEFAULTS = {
     SELECTED_EVENT_BOX_DRAWER: "Default Event box drawer",
     WINDOW_WIDTH: "900",
@@ -107,7 +108,8 @@ DEFAULTS = {
     FUZZY_ICON: "fuzzy.png",
     LOCKED_ICON: "locked.png",
     HYPERLINK_ICON: "hyperlink.png",
-    DATE_FORMAT: "yyyy-mm-dd"
+    DATE_FORMAT: "yyyy-mm-dd",
+    VERTCAL_SPACE_BETWEEN_EVENTS: "5",
 }
 # Some settings
 MAX_NBR_OF_RECENT_FILES_SAVED = 5
@@ -447,6 +449,13 @@ class Config(Observable):
         self.config_parser.set(DEFAULTSECT, DATE_FORMAT, date_format)
         self._notify()
     date_format = property(get_date_format, set_date_format)
+
+    def get_vertical_space_between_events(self):
+        return self.config_parser.getint(DEFAULTSECT, VERTCAL_SPACE_BETWEEN_EVENTS)
+
+    def set_vertical_space_between_events(self, value):
+        self.config_parser.set(DEFAULTSECT, VERTCAL_SPACE_BETWEEN_EVENTS, str(max(0, value)))
+        self._notify()
 
     def _toStr(self, value):
         try: