+++ source/timelinelib/wxgui/dialogs/preferences/controller.py
@@ -96,6 +96,9 @@ class PreferencesDialogController(Controller):
         self.config.set_hyperlink_icon(event.GetString())
         self.view.DisplayIcons()
 
+    def on_vertical_space_between_events_click(self, event):
+        self.config.set_vertical_space_between_events(self.view.GetVerticalSpaceBetweenEvents())
+
     def _set_initial_values(self):
         self.view.SetOpenRecentCheckboxValue(self.config.get_open_recent_at_startup())
         self.view.SetInertialScrollingCheckboxValue(self.config.get_use_inertial_scrolling())
@@ -115,6 +118,7 @@ class PreferencesDialogController(Controller):
         self.view.SetHyperlinkIcon(self.config.get_hyperlink_icon())
         self.view.SetCurrentDateFormat("%s: %s" % (_("Current"), self.config.date_format))
         self.view.DisplayIcons()
+        self.view.SetVerticalSpaceBetweenEvents(self.config.get_vertical_space_between_events())
 
     def _week_index(self, week):
         for (i, w) in self.weeks_map: